# perf(java): Refactor ThreadPoolFury to improve performance