<!--
**Thanks for contributing to Fury.**

**If this is your first time opening a PR on fury, you can refer to [CONTRIBUTING.md](https://github.com/apache/fury/blob/main/CONTRIBUTING.md).**

Contribution Checklist

    - The **Apache Fury (incubating)** community has restrictions on the naming of pr titles. You can also find instructions in [CONTRIBUTING.md](https://github.com/apache/fury/blob/main/CONTRIBUTING.md).

    - Fury has a strong focus on performance. If the PR you submit will have an impact on performance, please benchmark it first and provide the benchmark result here.
-->

## What does this PR do?

<!-- Describe the purpose of this PR. -->
After testing, it was observed that ​ThreadPoolFury​ experiences prolonged blocking during cold starts under high-concurrency scenarios. Analysis revealed that improper usage of locks in ​ClassLoaderFuryPooled​ was the root cause. This PR refactors the implementation of ​ClassLoaderFuryPooled​ by significantly reducing the granularity of locks, thereby drastically minimizing blocking time during cold starts.

## Related issues

<!--
Is there any related issue? Please attach here.

- #xxxx0
- #xxxx1
- #xxxx2
-->
[](url)https://github.com/apache/fury/issues/2087
## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an issue](https://github.com/apache/fury/issues/new/choose) describing the need to do so and update the document if necessary.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the PR will have an impact on performance, you can submit the PR first, and if it will have impact on performance, the code reviewer will explain it), be sure to attach a benchmark data here.
-->
CPU:9950X 
Class:org.apache.fury.benchmark.ThreadPoolFurySuite.java
old:
  Percentiles, ms/op:
      p(0.0000) =      0.001 ms/op
     p(50.0000) =      0.001 ms/op
     p(90.0000) =   1587.388 ms/op
     p(95.0000) =   1587.388 ms/op
     p(99.0000) =   1587.388 ms/op
     p(99.9000) =   1587.388 ms/op
     p(99.9900) =   1587.388 ms/op
     p(99.9990) =   1587.388 ms/op
     p(99.9999) =   1587.388 ms/op
    p(100.0000) =   1587.388 ms/op

new:
  Percentiles, ms/op:
      p(0.0000) =      0.001 ms/op
     p(50.0000) =      0.001 ms/op
     p(90.0000) =     62.746 ms/op
     p(95.0000) =     62.746 ms/op
     p(99.0000) =     62.746 ms/op
     p(99.9000) =     62.746 ms/op
     p(99.9900) =     62.746 ms/op
     p(99.9990) =     62.746 ms/op
     p(99.9999) =     62.746 ms/op
    p(100.0000) =     62.746 ms/op