@wuchaococo 2025-03-13
@moooonk 
当核心线程数不够时，调用addFury会把变量activeCacheNumber加上去，后面没有减回来，会导致当这个线程被回收后，后面核心线程数不够就走不到addFury逻辑，只能阻塞等其他线程释放了